+++ CHANGELOG.md
@@ -6,6 +6,15 @@
 - CVE-2022-1049: Pcs daemon was allowing expired accounts, and accounts with
   expired passwords to login when using PAM auth. ([huntr#220307],
   [rhbz#2068457])
+- Pcsd does not expose the server name in HTTP headers anymore ([rhbz#2059122])
+- Set `Strict-Transport-Security: max-age=63072000` HTTP header for all
+  responses ([rhbz#2097731])
+- Set HTTP headers to prevent caching everything except static files
+  ([rhbz#2097733])
+- Set HTTP headers to prevent sending referrer ([rhbz#2097732])
+- Set cookie option SameSite to Lax ([rhbz#2097730])
+- Set `Content-Security-Policy: frame-ancestors 'self'; default-src 'self'`
+  HTTP header for all responses ([rhbz#2097778])
 
 ### Added
 - Add support for fence\_mpath to `pcs stonith update-scsi-devices` command
@@ -36,8 +45,14 @@
 [rhbz#2058247]: https://bugzilla.redhat.com/show_bug.cgi?id=2058247
 [rhbz#2058251]: https://bugzilla.redhat.com/show_bug.cgi?id=2058251
 [rhbz#2058252]: https://bugzilla.redhat.com/show_bug.cgi?id=2058252
+[rhbz#2059122]: https://bugzilla.redhat.com/show_bug.cgi?id=2059122
 [rhbz#2068457]: https://bugzilla.redhat.com/show_bug.cgi?id=2068457
 [rhbz#2095695]: https://bugzilla.redhat.com/show_bug.cgi?id=2095695
+[rhbz#2097730]: https://bugzilla.redhat.com/show_bug.cgi?id=2097730
+[rhbz#2097731]: https://bugzilla.redhat.com/show_bug.cgi?id=2097731
+[rhbz#2097732]: https://bugzilla.redhat.com/show_bug.cgi?id=2097732
+[rhbz#2097733]: https://bugzilla.redhat.com/show_bug.cgi?id=2097733
+[rhbz#2097778]: https://bugzilla.redhat.com/show_bug.cgi?id=2097778
 [huntr#220307]: https://huntr.dev/bounties/7aa921fc-a568-4fd8-96f4-7cd826246aa5/